splitting files by logic
