+++ cmd/exporter.go
@@ -23,114 +23,6 @@ const (
 	userAgent = "gitlab-ci-pipelines-exporter"
 )
 
-// Client holds a GitLab client
-type Client struct {
-	*gitlab.Client
-	RateLimiter ratelimit.Limiter
-}
-
-var (
-	coverage = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_coverage",
-			Help: "Coverage of the most recent pipeline",
-		},
-		[]string{"project", "topics", "ref"},
-	)
-
-	lastRunDuration = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_last_run_duration_seconds",
-			Help: "Duration of last pipeline run",
-		},
-		[]string{"project", "topics", "ref"},
-	)
-
-	lastRunJobDuration = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_last_job_run_duration_seconds",
-			Help: "Duration of last job run",
-		},
-		[]string{"project", "topics", "ref", "stage", "job_name"},
-	)
-
-	lastRunJobStatus = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_last_job_run_status",
-			Help: "Status of the most recent job",
-		},
-		[]string{"project", "topics", "ref", "stage", "job_name", "status"},
-	)
-
-	lastRunJobArtifactSize = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_last_job_run_artifact_size",
-			Help: "Filesize of the most recent job artifacts",
-		},
-		[]string{"project", "topics", "ref", "stage", "job_name"},
-	)
-
-	timeSinceLastJobRun = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_time_since_last_job_run_seconds",
-			Help: "Elapsed time since most recent GitLab CI job run.",
-		},
-		[]string{"project", "topics", "ref", "stage", "job_name"},
-	)
-
-	jobRunCount = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "gitlab_ci_pipeline_job_run_count",
-			Help: "GitLab CI pipeline job run count",
-		},
-		[]string{"project", "topics", "ref", "stage", "job_name"},
-	)
-
-	lastRunID = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_last_run_id",
-			Help: "ID of the most recent pipeline",
-		},
-		[]string{"project", "topics", "ref"},
-	)
-
-	lastRunStatus = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_last_run_status",
-			Help: "Status of the most recent pipeline",
-		},
-		[]string{"project", "topics", "ref", "status"},
-	)
-
-	runCount = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "gitlab_ci_pipeline_run_count",
-			Help: "GitLab CI pipeline run count",
-		},
-		[]string{"project", "topics", "ref"},
-	)
-
-	timeSinceLastRun = prometheus.NewGaugeVec(
-		prometheus.GaugeOpts{
-			Name: "gitlab_ci_pipeline_time_since_last_run_seconds",
-			Help: "Elapsed time since most recent GitLab CI pipeline run.",
-		},
-		[]string{"project", "topics", "ref"},
-	)
-)
-
-var defaultMetrics = []prometheus.Collector{coverage, lastRunDuration, lastRunID, lastRunStatus, runCount, timeSinceLastJobRun, lastRunDuration, lastRunJobStatus, jobRunCount, timeSinceLastJobRun, lastRunJobArtifactSize}
-
-func newMetricsRegistry(log *log.Logger, metrics ...prometheus.Collector) *prometheus.Registry {
-	registry := prometheus.NewRegistry()
-	for _, m := range metrics {
-		if err := registry.Register(m); err != nil {
-			log.Fatalf("could not add provided metric '%v' to the Prometheus registry: %v", m, err)
-		}
-	}
-	return registry
-}
-
 // Run launches the exporter
 func Run(ctx *cli.Context) error {
 
@@ -191,7 +83,7 @@ func Run(ctx *cli.Context) error {
 	}
 
 	// Register the default metrics into a new registry
-	registry := newMetricsRegistry(log.StandardLogger(), defaultMetrics...)
+	registerMetricOn(registry, log.StandardLogger(), defaultMetrics...)
 
 	// Expose the registered registry via HTTP
 	mux := http.NewServeMux()
@@ -225,9 +117,9 @@ func Run(ctx *cli.Context) error {
 	return exit(nil, 0)
 }
 
-func metricsHandlerFor(metrics *prometheus.Registry, openMetricsEncoder bool) http.Handler {
-	return promhttp.HandlerFor(metrics, promhttp.HandlerOpts{
-		Registry:          metrics,
+func metricsHandlerFor(registry *prometheus.Registry, openMetricsEncoder bool) http.Handler {
+	return promhttp.HandlerFor(registry, promhttp.HandlerOpts{
+		Registry:          registry,
 		EnableOpenMetrics: openMetricsEncoder,
 	})
 }
